Change coordinates to double

+++ app/src/main/java/com/example/fieldsurfacearea/Point.kt
@@ -1,3 +1,3 @@
 package com.example.fieldsurfacearea
 
-class Point(val latitude: Float, val longitude: Float)
+class Point(val latitude: Double, val longitude: Double)
